requirements.txt
# Сегментация объектов пайки волноводов
# Python 3.10+

ultralytics>=8.1.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
PyYAML>=6.0
torch>=2.0.0
matplotlib>=3.7.0
